Refactor entire redirection mechanism to be handled with next.config.js redirects
