+++ packages/web/pages/dashboard/index.tsx
@@ -1,1 +1,8 @@
-export { default } from '../my-feed'
+/**
+ * This page is part of the deprecated URL pattern: `/dashboard/*` and should not be used.
+ * See the `redirects` config property in `next.config.js` for all redirect paths.ss
+ */
+
+export default function () {
+  return null
+}
