+++ packages/web/pages/dashboard/my-posts.tsx
@@ -1,14 +1,8 @@
-import { useRouter } from 'next/router'
-import { withApollo } from '@/lib/apollo'
+/**
+ * This page is part of the deprecated URL pattern: `/dashboard/*` and should not be used.
+ * See the `redirects` config property in `next.config.js` for all redirect paths.ss
+ */
 
-const OldMyPostsPage = () => {
-  const router = useRouter()
-  if (typeof window !== 'undefined') {
-    router.push({
-      pathname: `/my-posts`,
-    })
-  }
+export default function () {
   return null
 }
-
-export default withApollo(OldMyPostsPage)
